refact(front): agregando funcion put
