+++ front/src/pages/PrivatePages/hooks/useDelete.jsx
@@ -1,1 +1,5 @@
-ffc
+function useDelete() {
+
+}
+
+export default useDelete;
